devide classes into different packages and add some new interfaces
