+++ app/src/main/java/com/ligrim/tower_defense/GameStage.java
@@ -12,7 +12,20 @@ import java.util.LinkedList;
 import java.util.List;
 import android.graphics.Bitmap;
 import android.graphics.Canvas;
-import android.text.LoginFilter;
+
+import com.ligrim.tower_defense.base.BFS;
+import com.ligrim.tower_defense.base.Position;
+import com.ligrim.tower_defense.enemy.BossEnemy;
+import com.ligrim.tower_defense.enemy.Enemy;
+import com.ligrim.tower_defense.enemy.NormalEnemy;
+import com.ligrim.tower_defense.enemy.SmallerEnemy;
+import com.ligrim.tower_defense.enemy.TankerEnemy;
+import com.ligrim.tower_defense.tile.GameTile;
+import com.ligrim.tower_defense.tile.Mountain;
+import com.ligrim.tower_defense.tower.MachineGunTower;
+import com.ligrim.tower_defense.tower.NormalTower;
+import com.ligrim.tower_defense.tower.SniperTower;
+import com.ligrim.tower_defense.tower.Tower;
 
 import org.w3c.dom.Attr;
 import org.w3c.dom.Document;
