--- app/src/main/java/com/ligrim/tower_defense/SmallerEnemy.java
@@ -1,18 +1,0 @@
-package com.ligrim.tower_defense;
-
-import android.graphics.Canvas;
-
-import java.util.List;
-
-public class SmallerEnemy extends Enemy {
-
-    public SmallerEnemy(List<Position> route) {
-        super("enemy_smaller", route);
-        this.health = 150;
-        this.maxHealth = health;
-        this.speed = 180f / 60;
-        this.armor = 0;
-        this.prize = 1;
-        faded = false;
-    }
-}
