+++ app/src/main/java/com/ligrim/tower_defense/base/Attacker.java
@@ -1,0 +1,9 @@
+package com.ligrim.tower_defense.base;
+
+import com.ligrim.tower_defense.Bullet;
+
+public interface Attacker {
+
+    Bullet attack();
+
+}
