+++ app/src/main/java/com/ligrim/tower_defense/base/Destroyable.java
@@ -1,0 +1,9 @@
+package com.ligrim.tower_defense.base;
+
+public interface Destroyable {
+
+    boolean isDestroyed();
+
+    int disappear();
+
+}
