+++ app/src/main/java/com/ligrim/tower_defense/base/Moveable.java
@@ -1,0 +1,6 @@
+package com.ligrim.tower_defense.base;
+
+public interface Moveable {
+
+    void move();
+}
