+++ app/src/main/java/com/ligrim/tower_defense/base/Vulnerable.java
@@ -1,0 +1,8 @@
+package com.ligrim.tower_defense.base;
+
+public interface Vulnerable {
+
+    void reduceHealth(int damage);
+
+    boolean isDead();
+}
