+++ app/src/main/java/com/ligrim/tower_defense/enemy/SmallerEnemy.java
@@ -1,0 +1,18 @@
+package com.ligrim.tower_defense.enemy;
+
+import com.ligrim.tower_defense.base.Position;
+
+import java.util.List;
+
+public class SmallerEnemy extends Enemy {
+
+    public SmallerEnemy(List<Position> route) {
+        super("enemy_smaller", route);
+        this.health = 150;
+        this.maxHealth = health;
+        this.speed = 180f / 60;
+        this.armor = 0;
+        this.prize = 1;
+        faded = false;
+    }
+}
