+++ app/src/main/java/com/ligrim/tower_defense/tower/SniperTower.java
@@ -1,0 +1,16 @@
+package com.ligrim.tower_defense.tower;
+
+import com.ligrim.tower_defense.base.Position;
+
+public class SniperTower extends Tower {
+
+    public SniperTower(Position position) {
+        super("tower_sniper", position);
+        rateOfFire = 150f / 60;
+        range = 600;
+        damage = 1000;
+        lastShotTick = 0;
+        bulletSpeed = 15f;
+        price = 15;
+    }
+}
